[Installer] Extract run_prepare_command in own controller class
